+++ example/otel-collector/main.go
@@ -31,7 +31,7 @@ import (
 	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.18.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.24.0"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
@@ -162,9 +162,11 @@ func main() {
 }
 
 func connectDB() *sql.DB {
+	attrs := append(otelsql.AttributesFromDSN(mysqlDSN), semconv.DBSystemMySQL)
+
 	// Connect to database
 	db, err := otelsql.Open("mysql", mysqlDSN, otelsql.WithAttributes(
-		semconv.DBSystemMySQL,
+		attrs...,
 	))
 	if err != nil {
 		log.Fatal(err)
@@ -172,7 +174,7 @@ func connectDB() *sql.DB {
 
 	// Register DB stats to meter
 	err = otelsql.RegisterDBStatsMetrics(db, otelsql.WithAttributes(
-		semconv.DBSystemMySQL,
+		attrs...,
 	))
 	if err != nil {
 		log.Fatal(err)