+++ CHANGELOG.md
@@ -8,6 +8,43 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.38.0] - 2025-03-26
+
+### Added
+
+- `WithInstrumentErrorAttributesGetter` option to provide additional error-related attributes. (#440)
+
+### Changed
+
+- Upgrade OTel to `v1.35.0/v0.57.0`. (#437)
+
+### Removed
+
+- Drop support for Go `1.22`. (#447)
+
+## [0.37.0] - 2025-02-16
+
+### Added
+
+- `AttributesFromDSN` method to generate `server.address` and `server.port` attributes from a DSN. (#419)
+- Go 1.24 to supported versions. (#422)
+
+### Changed
+
+- Upgrade OTel to `v1.34.0/v0.56.0`. (#412)
+- Update the comment for the `WithAttributes` option to correctly describe the behavior on measurement creation. (#413)
+- Upgrade semantic conventions to `semconv/v1.24.0`. (#418)
+
+## [0.36.0] - 2024-12-18
+
+### Added
+
+- `DisableSkipErrMeasurement` option suppresses `driver.ErrSkip` as an error status in measurements if enabled. (#389)
+
+### Changed
+
+- Upgrade OTel to `v1.33.0/v0.55.0`. (#396)
+
 ## [0.35.0] - 2024-10-11
 
 ### Changed
@@ -387,7 +424,10 @@ It contains instrumentation for trace and depends on OTel `v0.18.0`.
 - Example code for a basic usage.
 - Apache-2.0 license.
 
-[Unreleased]: https://github.com/XSAM/otelsql/compare/v0.35.0...HEAD
+[Unreleased]: https://github.com/XSAM/otelsql/compare/v0.38.0...HEAD
+[0.38.0]: https://github.com/XSAM/otelsql/releases/tag/v0.38.0
+[0.37.0]: https://github.com/XSAM/otelsql/releases/tag/v0.37.0
+[0.36.0]: https://github.com/XSAM/otelsql/releases/tag/v0.36.0
 [0.35.0]: https://github.com/XSAM/otelsql/releases/tag/v0.35.0
 [0.34.0]: https://github.com/XSAM/otelsql/releases/tag/v0.34.0
 [0.33.0]: https://github.com/XSAM/otelsql/releases/tag/v0.33.0